requirements.txt
streamlit
python-dotenv
langchain
langchain-core
langchain_community
yfinance
pydantic
streamlit
python-dotenv
pandas
numpy
matplotlib
pickle-mixin
ipython
llama-index
pyvis
azure-storage-blob
azure-identity
openai
azure-core
azure-ai-formrecognizer
requests
ipdb